requirements.txt
# Core AI and ML
torch>=2.0.0
torchvision==0.15.2
transformers>=4.30.0
numpy>=1.21.0
scipy==1.10.1

# Computer Vision
opencv-python==4.8.0.74
mediapipe>=0.10.0
face-recognition==1.3.0
dlib==19.24.0

# Natural Language Processing
nltk==3.8.1
spacy>=3.5.0

# Audio Processing
sounddevice==0.4.6
librosa==0.10.0
pyaudio==0.2.13
SpeechRecognition==3.10.0

# Web and API
fastapi==0.100.0
uvicorn==0.22.0
websockets==11.0.3
python-dotenv>=1.0.0

# Utilities
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.1
tqdm==4.65.0
tensorflow>=2.12.0
requests>=2.31.0
pyttsx3>=2.90 